fix: streamline type imports and definitions for consistency across board features
