+++ src/features/board/index.ts
@@ -1,4 +1,5 @@
 export { BoardView } from "./BoardView";
+export type { BoardRouteParams } from "./navigation/types";
 export {
   getBoardSets,
   importBoardFromUrl,
@@ -6,6 +7,5 @@ export {
 } from "./storage/board-sets-store";
 export type { BoardSetRecord } from "./storage/boards-db";
 export { useBoardSets } from "./storage/useBoardSets";
-export type { BoardRouteParams } from "./types";
 export { useBoard } from "./useBoard";
 export { useImportBoardFiles } from "./useImportBoardFiles";
