+++ src/features/board/navigation/types.ts
@@ -1,0 +1,5 @@
+export interface BoardRouteParams {
+  [key: string]: string;
+  setId: string;
+  boardId: string;
+}
