+++ src/features/board/suggestions/SuggestionBar.tsx
@@ -1,8 +1,8 @@
 import Box from "@mui/material/Box";
 import Chip from "@mui/material/Chip";
 import Stack from "@mui/material/Stack";
-import type { SuggestionTone } from "../types";
 import { ToneSelector } from "./ToneSelector";
+import type { SuggestionTone } from "./types";
 
 export interface SuggestionBarProps {
   suggestions: string[];
