+++ src/features/board/suggestions/types.ts
@@ -1,0 +1,2 @@
+// RewriterTone is a global ambient type from @types/dom-chromium-ai.
+export type SuggestionTone = RewriterTone;
