+++ src/features/board/types.ts
@@ -1,37 +1,18 @@
-export interface BoardRouteParams {
-  [key: string]: string;
-  setId: string;
-  boardId: string;
+export interface Board {
+  id: string;
+  name?: string;
+  locale?: string;
+  grid: BoardGrid;
+  buttons: BoardButton[];
+  descriptionHTML?: string;
+  license?: BoardLicense;
+  strings?: BoardStrings;
 }
-
-export type SuggestionTone = RewriterTone;
-
-type SpecialtyAction = ":space" | ":clear" | ":home" | ":speak" | ":backspace";
-type SpellingAction = `+${string}`;
-
-export type BoardAction = SpecialtyAction | SpellingAction;
 
 export interface BoardGrid {
   rows: number;
   columns: number;
   order?: (string | null)[][];
-}
-
-export interface BoardLicense {
-  type: string;
-  authorName?: string;
-  authorEmail?: string;
-  authorUrl?: string;
-  sourceUrl?: string;
-  copyrightNoticeUrl?: string;
-}
-
-export interface BoardLink {
-  id?: string;
-  name?: string;
-  url?: string;
-  path?: string;
-  dataUrl?: string;
 }
 
 export interface BoardButton {
@@ -46,15 +27,27 @@ export interface BoardButton {
   loadBoard?: BoardLink;
 }
 
+export interface BoardLicense {
+  type: string;
+  authorName?: string;
+  authorEmail?: string;
+  authorUrl?: string;
+  sourceUrl?: string;
+  copyrightNoticeUrl?: string;
+}
+
 export type BoardStrings = Record<string, Record<string, string>>;
 
-export interface Board {
-  id: string;
+export interface BoardLink {
+  id?: string;
   name?: string;
-  locale?: string;
-  grid: BoardGrid;
-  buttons: BoardButton[];
-  descriptionHTML?: string;
-  license?: BoardLicense;
-  strings?: BoardStrings;
+  url?: string;
+  path?: string;
+  dataUrl?: string;
 }
+
+export type BoardAction = SpecialtyAction | SpellingAction;
+
+type SpecialtyAction = ":space" | ":clear" | ":home" | ":speak" | ":backspace";
+
+type SpellingAction = `+${string}`;
